Update extension verification instructions (#10)
In this step in the guide there are no plugins that requires the
extensions. This means that it will not be loaded automatically, and will
not show up in the extension list.

To verify that the extensions are properly installed, we can instead ask the
user to load the extensions. If they are not properly installed, they will not load.
If the extensions are properly installed, they will show a success message.